+++ src/database/MapCards.java
@@ -1,6 +1,7 @@
 package database;
 
 import javax.xml.bind.annotation.XmlRootElement;
+import java.math.BigInteger;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -8,14 +9,14 @@ import java.util.Map;
 @XmlRootElement
 public class MapCards
 {
-    private Map<Long, Card> mapCards;
+    private Map<BigInteger, Card> mapCards;
 
-    public Map<Long, Card> getMapCards()
+    public Map<BigInteger, Card> getMapCards()
     {
         return mapCards;
     }
 
-    public void setMapCards(Map<Long, Card> mapCards)
+    public void setMapCards(Map<BigInteger, Card> mapCards)
     {
         this.mapCards = mapCards;
     }
